refactor(shared): ポート関連ユーティリティの独立化と重複排除
Issue #184: shared/utils/port.ts を新設し、validatePort および
getPortFromUrl を移動。エイリアスを使用しない実装にすることで
Vite 設定フェーズでも安全に利用可能に。
これにより、vite.config.ts 内の重複ロジックを削除。
既存の shared/utils/url.ts からは re-export を行い後方互換性を維持。
